+++ src/enums/documentTypes.ts
@@ -1,0 +1,3 @@
+export enum DocumentClasses {
+  CORP = 'CORP'
+}
