+++ src/utils/AxiosInstance.ts
@@ -11,7 +11,6 @@ const instance = axios.create()
 instance.interceptors.request.use(
   request => {
     // don't add bearer token for Minio requests
-    console.log(sessionStorage.getItem)
     if (request.url?.startsWith('https://minio') || request.url.startsWith(sessionStorage.getItem('DRS_API_URL'))) {
       return request
     }
